Refactor service implementation
Remove ECS constainer defintion schema types that are now available in @pulumi/aws.

Factor docker build/push code out into a seperate module.
