+++ aws/tsconfig.json
@@ -26,6 +26,7 @@
         "service.ts",
         "unhandledError.ts",
         "utils.ts",
+        "docker.ts",
 
         "config/index.ts",
